db/requirements.txt
google.generativeai
pymongo
sentence-transformers
chromadb
tqdm
